rename algorithm to solvers
